bug fix for number exponent >= 256

+++ core/src/main/java/com/alibaba/fastjson2/JSONReaderUTF8.java
@@ -1875,7 +1875,7 @@ class JSONReaderUTF8
                 expValue = -expValue;
             }
 
-            this.exponent = (byte) expValue;
+            this.exponent = (short) expValue;
             valueType = JSON_TYPE_DEC;
         }
 
@@ -2113,7 +2113,7 @@ class JSONReaderUTF8
                 expValue = -expValue;
             }
 
-            this.exponent = (byte) expValue;
+            this.exponent = (short) expValue;
             valueType = JSON_TYPE_DEC;
         }
 
@@ -3441,7 +3441,7 @@ class JSONReaderUTF8
                 expValue = -expValue;
             }
 
-            this.exponent = (byte) expValue;
+            this.exponent = (short) expValue;
             valueType = JSON_TYPE_DEC;
         }
 
@@ -3655,7 +3655,7 @@ class JSONReaderUTF8
                 expValue = -expValue;
             }
 
-            this.exponent = (byte) expValue;
+            this.exponent = (short) expValue;
             valueType = JSON_TYPE_DEC;
         }
