+++ core/src/test/java/com/alibaba/fastjson2/JSONReaderStr.java
@@ -1965,7 +1965,7 @@ final class JSONReaderStr
                 expValue = -expValue;
             }
 
-            this.exponent = (byte) expValue;
+            this.exponent = (short) expValue;
             valueType = JSON_TYPE_DEC;
         }
 
@@ -2203,7 +2203,7 @@ final class JSONReaderStr
                 expValue = -expValue;
             }
 
-            this.exponent = (byte) expValue;
+            this.exponent = (short) expValue;
             valueType = JSON_TYPE_DEC;
         }
 
@@ -3021,7 +3021,7 @@ final class JSONReaderStr
                 expValue = -expValue;
             }
 
-            this.exponent = (byte) expValue;
+            this.exponent = (short) expValue;
             valueType = JSON_TYPE_DEC;
         }
